requirements.txt
# Core data science
jupyterlab==4.0.9
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4

# Machine learning
scikit-learn==1.3.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Web framework
streamlit==1.29.0

# NLP
gensim==4.3.2
nltk==3.8.1

# Utilities
python-dotenv==1.0.0
pyarrow==14.0.1
joblib==1.3.2
pyyaml==6.0.1
tqdm==4.66.1

# Optional image processing
scikit-image==0.22.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
